Add pre-populated initial data to BusinessInformationView
This required refactoring the Company model into two models. One for our internal representation of a Company entity and a second to register each dnb-service get_company response. The get_company response has a FK to the company model. This allows us to save the dnb-company data recorded at the time of the grant application and it allows us to use that data along during any point in the SME application form views.