Revert back to use circleCI python orb to speed up build/test time
Note: Cannot generate ERDs using the python orb because graphviz needs to be installed on the machine. We could install graphviz in a circleci build step (via apt-get install graphviz) however that would be difficult and complex to cache. So we generate the diagram in a separate job for now